fix: add custom cookies config to middleware
