+++ apps/ai-dial-admin/src/middleware.ts
@@ -1,7 +1,9 @@
-import { getIsEnableAuthToggle } from '@/src/utils/env/get-auth-toggle';
 import withAuth from 'next-auth/middleware';
 import { createI18nMiddleware } from 'next-international/middleware';
 import { NextRequest } from 'next/server';
+
+import { cookies } from '@/src/utils/auth/auth-cookies';
+import { getIsEnableAuthToggle } from '@/src/utils/env/get-auth-toggle';
 
 const I18nMiddleware = createI18nMiddleware({
   defaultLocale: 'en',
@@ -16,6 +18,8 @@ async function middlewareFn(req: NextRequest) {
   return I18nMiddleware(req);
 }
 
-const middleware = !getIsEnableAuthToggle() ? middlewareFn : withAuth(middlewareFn);
+const authMiddleware = withAuth(middlewareFn, { cookies });
+
+const middleware = !getIsEnableAuthToggle() ? middlewareFn : authMiddleware;
 
 export default middleware;
